osd: detect multi-actuator HDDs
If we are running on a device that has multiple LUNs, then note that in
our CRUSH hierarchy location.

Notably, if we pass a key=value location field that the mon doesn't
understand, it is ignored, so this shouldn't create any errors on
upgrade.  However, an upgrade cluster won't know what the 'device' crush
type means (yet).

Signed-off-by: Sage Weil <sage@newdream.net>