osd: Handle mclock queue's return value to track client's status
The original changes are adapted from the commits of
PR: https://github.com/ceph/ceph/pull/20235 to work with the
current implementation of mClock scheduler.

Change description:
Incorporate qos_cost and qos_phase into the mclock queue's return value
upon dequeue operation. These parameters are passed back via MOSDOPReply.
The cost and phase parameters are fed to the service tracker object by
the Objecter as part of handle_osd_op_reply(). This ensures that the
delta and rho QoS parameters are calculated for the next op from the
client.

Modify the existing unit test infrastructure to support specifying client
QoS request parameters. Add a new test that specifies the QoS request
params and verify that the items dequeued are based on the delta and rho
parameters specified in the request. Note that this is not an end-to-end
test and the unit test simply performs enqueue and dequeue operations on
the mClock queue. The dequeue results are an indication of the correctness
of the mClock algorithm.

The original commits from where the changes are adapted, their authors and
description are mentioned below.

1. bbae600d811642e084a1ce732037e04c7bba575f
   Author(s): Byungsu Park <bspark8@sk.com>
   Description:  osd: Handle the mclock queue's return value to track the
                 each dmclock queue

Adapted by:
Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>

Original Author(s):
Signed-off-by: Byungsu Park <bspark8@sk.com>