doc/cephadm: reorg cephadm front matter
My improved understanding of the Ceph ecosystem
has made it possible for me to see places in the
documentation where we can make improvements that
go beyond the mere correction of English syntax
and ensuring that the verbs agree with their
subjects. This PR is the first in what will likely
be a series of PRs that explain the parts of cephadm
as though the reader has absolutely no prior knowledge
of what cephadm is or what it does.

With this PR, I want to ask if we really need the sentence
that mentions "day 2 commands", because this seems like
marketing language to me. Is it possible to replace this
with something like "post-installation maintenance operations
and the commands that make those operations possible", or
is does the Ceph developer community have a strong desire to
keep the phrase "day 2 operations"?

inb4: I'm not trying to start a fight or anything like that.
I'm just wondering if there's a good reason to use the phrase
"day 2 operations".

Signed-off-by: Zac Dover <zac.dover@gmail.com>