mgr/dashboard: add unit test for telemetry replacer method
This unit test checks that the "replacer" method in telemetry.component.ts works as it should. The replacer method takes the telemetry report and changes the ranges and values of the 'osd_perf_histograms' field from arrays to strings, thereby making the report more readable in the Dashboard Telemetry Preview.

This unit test needs improvement since it currently uses a test report rather than the real one.

Signed-off-by: Laura Flores <lflores@redhat.com>