rpm: enable system_pmdk bcond for SUSE builds
5695349990db41778def6facb6995cb575a5b0b7 disabled the system_pmdk bcond for all
build targets based on the fact that pmdk 1.10 was not available on any of them.

Now that openSUSE Tumbleweed ships pmdk 1.11, we re-enable the system_pmdk bcond
to fix the master build for openSUSE Tumbleweed.

Since openSUSE Tumbleweed is the *only* SUSE build target master supports, there
is no need for greater granularity in the distro conditional here.

Signed-off-by: Nathan Cutler <ncutler@suse.com>