seastore: add reserved addr for RBM if small fresh extent
According to RBM design, small data blocks need to be appended
to CBJournal first---the small case should be inline
(writing to CBJournal). In addition, in this case, the extent
will be moved to RBM space eventually.
When the extent is moved from CBJournal to RBM, RBM might have
not enough space to store the extent. To avoid this, we need to reserve
the space for the extent before appending the extent to CBJournal.

But, in current EPM implementation,
there is no way to reserve RBM space if write strategy is
inline.

Therefore, this commit reserves RBM space by using rbm.alloc_extent()
and adds reserve_ool_addr to indicate where the extent is located in RBM
when the extent is fresh extent.

Signed-off-by: Myoungwon Oh <myoungwon.oh@samsung.com>