mgr/cephadm: put list_servers output in debug logs
Over time, seen a number of cases where daemons are being
marked stray for unknown reasons and it is always difficult
to debug due to the difficulty of finding what list_servers
is outputting. I think it would help to include this in the
debug logs. That way we have that side of the info available
and can easily compare it to daemon/host info cephadm is storing.

Signed-off-by: Adam King <adking@redhat.com>
(cherry picked from commit bdc802354554ea9c5e7c0377e3fe695f719921a7)