os/bluestore: Set min_alloc_size to optimal io size
Block devices may report an "optimal_io_size" that is different than the
typical 4KiB. To optimize BlueStore for this io size, the allocator
needs to set its min_alloc_size to this optimal_io_size. This PR adds
the discovery of the optimal_io_size for a block device and an option
to use the optimal_io_size as the min_alloc_size for the bluestore allocator.

Older devices may report an optimal_io_size of 0 and if that is the
case, the default config min_alloc_size is used.

Signed-off-by: Curt Bruns <curt.e.bruns@gmail.com>