doc/cephadm: update typo in install.rst
This PR changes the string

   use "--log-to-file" option during bootstrap

to

   use the "--log-to-file" option during bootstrap

The way that it was written makes it sound like it was
written by a native Russian speaker. But it was written
by me, a native English speaker, so this is on me.

Signed-off-by: Zac Dover <zac.dover@gmail.com>
(cherry picked from commit d5f6d8ef33413906c8a53a1196ad37a1e1e80289)