tools/cephfs-data-scan: Recover symlinks as symlinks
Problem:
The 'cephfs-data-scan' recovers the symlink as regular file
during disaster recovery of metadata pool from data pool.

Solution:
The symlink target is stored on the first data object
along with backtrace when the 'mds_symlink_recovery'
option is enabled. The option is enabled by default.
This patch adds the ability to the 'cephfs-data-scan'
tool recover symlink as symlink.

Signed-off-by: Kotresh HR <khiremat@redhat.com>
Fixes: https://tracker.ceph.com/issues/46166