librbd/BlockGuard: need increase overlap range when overlap  occur.
We use BlockGuard to avoid reorder when overlap occur. But the following
case don't work:
a: write(3,5)
b: write(4,9)
c: write(8, 11).
b overlap w/ a, so it append in a. c overlap w/ b.
In fact, c shuld append b. But currently, it can't change overlap range
when b overlap a occur. The c can't be detained and issue before b.
We should change overlap from (3,5) to (3, 9) when b overlap a.

Currently, we use boost::intrusive::set to store extent. But for the
following writes:
a(3,5), b(8,11), c(4,9).
When insert c(4,9), it overlap w/ a(3,5). When remove a and insert
a'(3,9). But a'(3,9) overlap w/ b(8,11). If we use container set,
we can't insert a'. So we change set to multiset to fix.
But when erase a, we should check whether containe a, etc a' cotaine a
but b can't contair a. So we should erase a' rather b.

Signed-off-by: Jianpeng Ma <jianpeng.ma@intel.com>