mgr/telemetry: modify stats_per_pool
There is a much easier way to collect stats_per_pool than the current implementation. Fetching 'pg_dump' from the mgr module already provides a field called "pool_stats" that is the same as aggregated pg stats, which was the implementation up until this commit.

All in all, this solution should provide the information we want, with a much cleaner implementation.

Signed-off-by: Laura Flores <lflores@redhat.com>

Backport Message: In the case that this commit is backported, it is important to note that the commits in PR #42569 should be backported first, as the implementation of "get_stat_sum_per_pool()" in #42569 precedes the removal of it here.

Signed-off-by: Laura Flores <lflores@redhat.com>
(cherry picked from commit 88e039231fac6d5b7ab8584933a13f022b121c23)

 Conflicts:
	src/pybind/mgr/telemetry/module.py

Had to add in mempool stats and remove stats_sum_per_pool since I am cherry-picking commits
that haven't been merged to master yet.