librbd/crypto: don't decrypt image uninitialized data
By convention, librbd returns zeros when reading uninitialized image data.
When using encryption, this convention is not guaranteed.
Specifically, this can happen when non-sparse reads are issued to the OSDs,
such in the case where the reads are smaller than rbd_sparse_read_threshold_bytes.
To overcome this, this commit adds a check to bypass decryption in case
the data read from the OSD is all zeros.

Fixes: https://tracker.ceph.com/issues/53417
Signed-off-by: Or Ozeri <oro@il.ibm.com>