orchestra/cluster: run parallel and wait
Make cluster.run able to execute command in parallel on all
nodes and wait for completion.

In order to run a command in parallel on all nodes of the
cluster, instead of combination:

  run.wait(cluster.run(wait=False, args=...))

this patch makes it possible to do the same just by:

  cluster.run(parallel=True, args=...)

the wait arguments defaults to True in this case.

Also the behavior of the method is not changed
when wait set to False (like in the first example),
so it is supposed to be backward compatible.

Signed-off-by: Kyr Shatskyy <kyrylo.shatskyy@suse.com>