msg: Fix Windows IPv6 support
The Windows AF_INET6 definition doesn't match the Linux one, thus
sockaddr structures must be converted when being sent over the wire.
Otherwise, Linux hosts will not be able to recognize the address
family and reject Windows client connections:

  handle_client_ident peer is trying to reach
  v2:(unrecognized address family 23)/0 which is not us

Note that this isn't the case with AF_INET, which has the same
definition on both platforms.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>