ceph-volume: adjust arguments for 'ceph-volume raw activate'
Take a list of devices, so that we can selectively activate a raw osd
with db/wal.

Remove the argument type kludge introduced in 2c228a9a409176c0f1679f176443fd3ead219c7a
since it is no longer needed.

Note that we're making this change because (1) it allows db/wal and (2)
because there are no known users of 'raw activate'.  The only known user
is via 'ceph-volume activate' and we've fixed that caller in this commit.

Signed-off-by: Sage Weil <sage@newdream.net>
(cherry picked from commit afd8be7eac5e996c3bd07656601a4534053e2516)

Conflicts:
	src/ceph-volume/ceph_volume/devices/raw/list.py