rgw: "reshard cancel" errors with "invalid argument"
When `radosgw-admin reshard cancel ...` was used on a bucket that is
not currently undergoing resharding, it would error out.

Now it will then try to unschedule any reshards for the bucket in the
so-called "reshard log". It will error out if there is neither an
underway resharding nor a scheduled resharding.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>
(cherry picked from commit cacddbb83d23fe6ba2d42e1497d0ff49fb880154)