mds/MDLog: use a larger segment size if the event max size is too large
In some case the ESubtreeMap event's size will be very large, which could
reach up to 4MB or larger. Then it's very possibly there will be many new
created LogSegments only contain two events, the first is ESubtreeMap event
and the second is an any other type event.

In this case the network bandwith usage maybe very high, which is mainly
writing the ESubtreeMap events to metatdata pool.

Add the "mds_log_segment_size_factor" option, which is the ratio of maximum
of MDLog's segment size and maximum size of every 10000 events at which
the MDLog will start a new LogSegment immediately if the maximum size of
the 10000 events is larger than half of the segment size. Set a value less
than or equal to 1 will disable it. This will make sure that there won't be
to many big events in the network traffic.

Fixes: https://tracker.ceph.com/issues/53542
Fixes: https://tracker.ceph.com/issues/53623
Signed-off-by: Xiubo Li <xiubli@redhat.com>