mds: add mds_log_max_average_segment_size_factor option support
In heavy load case the MDLog's actual segment size is very probably
larger than the expected segment size, which is Rados object's size.
Such as it could be 100 * Rados object's size, so in this case in
the worst case the journal logs could occupy around 200 * 4MB * 128,
which is 100GB or even larger.

Here we will limit it to mds_log_max_average_segment_size_factor as
default, if the actual average size of the segments is larger than
the factor * expected segment size, try to trim the segments.

Fixes: https://tracker.ceph.com/issues/40002
Signed-off-by: Xiubo Li <xiubli@redhat.com>