osdmaptool: explicit output filename cmd arg
This is a breaking change. The positional argument is not used as output
filename anymore, also removed --save. Now we must pass -o to save an osdmap.

This is more intuitive and consistent with crushtool.

This enables --createsimple/--import_crush to be used without save the osdmap
to a file. Or just decode and encode a map without changes.

Signed-off-by: 胡玮文 <huww98@outlook.com>