global: cleanup stall comments
Since aa22f2e6134 (global/global_init: fix stdout/stderr/stdin closing for daemonization),
we no longer use the trick of close-then-re-open. Instead, we use dup2,
which does not rely on the "old unix behavior that the file descriptor
that gets assigned is the lowest available one." So these comment is
out-of-date and inaccurate now.

Signed-off-by: 胡玮文 <huww98@outlook.com>