rgwlc:  harmonize lifecycle notifications with the others
After working with notification configurations, it seemed to make
more sense to organize the different lifecycle notification sub-types
along the same lines as the official AWS ones.  So, there are now
ObjectExpirationCurrent and ObjectExpirationNoncurrent types, and
an ObjectExpiration wildcard.  Similarly, ObjectTransition
contains Current and Noncurrent sub-types.

Note that I have defined an ObjectExpirationAbortMPU notification
sub-type, but do not currently generate it--this is to avoid changing
the rgw::sal interface, but most likely we would support this in
future.

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>