rgwlc:  introduce optional_cookie to LCEntry and cognate types
The optional_cookie tracks value of RGWLC::cookie, which is a random
value set at server/instance boot time.  IOW, this is a way to detect
whether the current, running radosgw instance is the one which last
updated a given LC entry.  The intended use is to robustify detection
of stale LC entries.

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>