mount.ceph: fix the handling of new-syntax device names
With the new mount device syntax, the mount helper will end up
prepending "client." to whatever device string is passed down. If you
then go to recreate that mount later from the info in /proc/mounts,
you'll end up getting back an error because it'll try to prepend
"client." again.

There is no reason to send a fully-qualified principal name down to the
kernel since it can only use client.* principals anyway. Fix the mount
helper to track the unqualified name internally and only fully-qualify
it when we're scraping the config for info.

Fixes: https://tracker.ceph.com/issues/53765
Signed-off-by: Jeff Layton <jlayton@redhat.com>