src/rados: rewrite bluestore-cache_autotune_interv
This commit rewrites one sentence in the "Bluestore
Config Reference". The sentence does not itself appear
in the directory structure under /docs/rados/, but is
instead in /src/common/options/global.yaml.in and is
included in the Bluestore Config Reference by means of
a confval directive.

The rewritten sentence has been rewritten from the passive
voice to the active voice, and a "how"-formation has been
removed. The removal of the "how"-formation, in my opinion,
makes the sentence go down much smoother for the first-time
reader.

Signed-off-by: Zac Dover <zac.dover@gmail.com>