README.md: Update README.md to add link to tracker.ceph.com
I searched the existing documentation for a link to the tracker and had a hard time finding it. Other folks like myself might like it so prominently displayed.
Added a link to https://tracker.ceph.com/projects/ceph as https://tracker.ceph.com/ is basically blank and not totally intuitive how to get to issues.

Signed-off-by: John Bent <john.bent@seagate.com>